swemi

+++ src/tx.c
@@ -1,35 +1,31 @@
 #include "la.h"
-#include "tx.h"
-#include "mo.h"
-#include "str.h"
 #include <string.h>
 
-static long tx_mo_n(la, FILE*, ob);
+static ssize_t tx_mo_n(la, FILE*, ob);
 
 // s-expression writer
-long la_tx(la v, FILE *o, ob x) {
+ssize_t la_tx(la v, FILE *o, ob x) {
   if (nump(x)) return fprintf(o, "%ld", getnum(x));
   if (G(x) == disp) return ((mtbl) GF(x))->emit(v, o, x);
   if (primp((mo) x)) return
     fprintf(o, "\\%s", ((struct la_prim*)x)->nom);
   return tx_mo_n(v, o, hnom(v, (mo) x)); }
 
-long fputstr(FILE *o, str s) {
-  long i = 0, r = s->len;
+ssize_t fputstr(FILE *o, str s) {
+  ssize_t i = 0, r = s->len;
   while (i < r) if (fputc(s->text[i++], o) == EOF) return -1;
   return r; }
 
 // FIXME this is really weird
 // print a function name
-#include "two.h"
-static NoInline long tx_mo_n(la v, FILE *o, ob x) {
+static NoInline ssize_t tx_mo_n(la v, FILE *o, ob x) {
   if (symp(x)) {
     if (fputc('\\', o) == EOF) return -1;
-    long r = la_tx(v, o, x);
+    ssize_t r = la_tx(v, o, x);
     return r < 0 ? r : r + 1; }
   if (!twop(x))
     return fputc('\\', o) == EOF ? -1 : 1;
-  long r = 0, a;
+  ssize_t r = 0, a;
   // FIXME this is weird
   if (symp(A(x)) || twop(A(x))) {
     a = tx_mo_n(v, o, A(x));
@@ -41,7 +37,6 @@ static NoInline long tx_mo_n(la v, FILE *o, ob x) {
     r += a; }
   return r; }
 
-#include "vm.h"
 Vm(tx_f) {
   size_t i = 0, l = fp->argc;
   if (l) {
